Refactoring TUI
