+++ client/internal/tui/auth/state.go
@@ -1,0 +1,12 @@
+package auth
+
+type state int
+
+const (
+	stateInit state = iota
+	stateAskUsername
+	stateAskPassword
+	stateInProgress
+	stateSuccess
+	stateError
+)
